Conflicts:
	FilmOverflow.WebUI/App_Data/aspnet-FilmOverflow.WebUI-20140814042149.mdf
	FilmOverflow.WebUI/App_Data/aspnet-FilmOverflow.WebUI-20140814042149_log.ldf
	FilmOverflow.WebUI/FilmOverflow.WebUI.csproj

+++ FilmOverflow.DAL/Models/Seat.cs
@@ -15,6 +15,6 @@ namespace FilmOverflow.DAL.Models
 		public bool IsProcessing { get; set; }
 
 		public virtual Hall Hall { get; set; }
-		public int HallId { get; set; }
+		public long HallId { get; set; }
 	}
 }
